Create move.js, use move.* names, move code there
